requirements.txt
numpy==1.23.5
pandas==1.5.1
torch==1.10.2
torchvision==0.11.3
pillow==8.4.0
